+++ tests/ops/service/saas_request/test_saas_request_override_factory.py
@@ -12,6 +12,7 @@ from fides.api.models.consent_automation import ConsentableItem
 from fides.api.models.policy import Policy
 from fides.api.models.privacy_notice import UserConsentPreference
 from fides.api.models.privacy_request import PrivacyRequest
+from fides.api.schemas.consentable_item import ConsentWebhookResult
 from fides.api.service.connectors.saas.authenticated_client import AuthenticatedClient
 from fides.api.service.saas_request.saas_request_override_factory import (
     SaaSRequestOverrideFactory,
@@ -99,6 +100,19 @@ def valid_consent_update_override(
     A sample override function for consent update requests with a valid function signature
     """
     return True
+
+
+def valid_process_consent_webhook_override(
+    client: AuthenticatedClient,
+    secrets: Dict[str, Any],
+    payload: Any,
+    notice_id_to_preference_map: Dict[str, UserConsentPreference],
+    consentable_items: List[ConsentableItem],
+) -> ConsentWebhookResult:
+    """
+    A sample override function for process consent webhook requests with a valid function signature
+    """
+    return ConsentWebhookResult()
 
 
 @pytest.mark.unit_saas
@@ -207,6 +221,26 @@ class TestSaasRequestOverrideFactory:
         register(f_id, SaaSRequestType.UPDATE_CONSENT)(valid_consent_update_override)
         assert valid_consent_update_override == SaaSRequestOverrideFactory.get_override(
             f_id, SaaSRequestType.UPDATE_CONSENT
+        )
+
+        with pytest.raises(NoSuchSaaSRequestOverrideException) as exc:
+            SaaSRequestOverrideFactory.get_override(f_id, SaaSRequestType.READ)
+        assert f"Custom SaaS override '{f_id}' does not exist." in str(exc.value)
+
+    def test_register_process_consent_webhook_override(self):
+        """
+        Test registering a valid `process_consent_webhook` override function
+        """
+
+        f_id = uuid()
+        register(f_id, SaaSRequestType.PROCESS_CONSENT_WEBHOOK)(
+            valid_process_consent_webhook_override
+        )
+        assert (
+            valid_process_consent_webhook_override
+            == SaaSRequestOverrideFactory.get_override(
+                f_id, SaaSRequestType.PROCESS_CONSENT_WEBHOOK
+            )
         )
 
         with pytest.raises(NoSuchSaaSRequestOverrideException) as exc: